Changed as per comments 2
