+++ src/main/java/Davydova/files/Address.java
@@ -1,5 +1,4 @@
 import java.util.ArrayList;
-import java.util.List;
 
 /**
  * This is the class for creating Collections
@@ -15,12 +14,6 @@ public class Address {
     private ArrayList<Address> addressList;
 
     /**
-     * Дефолтный конструктор класса Address
-     */
-    public Address() {
-    }
-
-    /**
      * Конструктор класса Address в котором инициализируются поля city, street, building, apartment
      * @param city
      * @param street
@@ -32,24 +25,6 @@ public class Address {
         this.street = street;
         this.building = building;
         this.apartment = apartment;
-    }
-
-    /**
-     * Получить коллекцию ArrayList класса Address
-     * @return возвращает коллекцию ArrayList класса Address
-     */
-    public ArrayList<Address> getAddressList() {
-        ArrayList<Address> addressList = new ArrayList<>();
-
-        addressList.add(new Address("Сочи", "Цветочная", 2, 11));
-        addressList.add(new Address("Тюмень", "Театральная", 65, 85));
-        addressList.add(new Address("Таганрог", "Ленина", 22, 51));
-        addressList.add(new Address("Москва", "Ленина", 56, 72));
-        addressList.add(new Address("Омск", "Цветочная", 5, 9));
-        addressList.add(new Address("Тольятти", "Портовая", 8, 10));
-        addressList.add(new Address("Оренбург", "Северная", 64, 2));
-
-        return addressList;
     }
 
     /**
